api/requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==2.0
Flask-Testing
SQLAlchemy-migrate